# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestRequest": {
        "additionalProperties": false,
        "properties": {
          "documents": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Documents",
            "type": "array"
          },
          "workspace_id": {
            "default": "default",
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "documents"
        ],
        "title": "IngestRequest",
        "type": "object"
      },
      "IngestResponse": {
        "additionalProperties": false,
        "properties": {
          "ingested": {
            "items": {
              "$ref": "#/components/schemas/Ingested"
            },
            "title": "Ingested",
            "type": "array"
          },
          "skipped": {
            "items": {
              "type": "string"
            },
            "title": "Skipped",
            "type": "array"
          },
          "store": {
            "title": "Store",
            "type": "string"
          }
        },
        "required": [
          "store"
        ],
        "title": "IngestResponse",
        "type": "object"
      },
      "Ingested": {
        "additionalProperties": false,
        "properties": {
          "chunks": {
            "default": 0,
            "title": "Chunks",
            "type": "integer"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "store": {
            "default": "",
            "title": "Store",
            "type": "string"
          },
          "unchanged": {
            "default": false,
            "title": "Unchanged",
            "type": "boolean"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "document_id"
        ],
        "title": "Ingested",
        "type": "object"
      },
...
      "RetrievalHit": {
        "additionalProperties": false,
        "description": "One passage, citable, with the score that ranked it.",
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "source",
          "document_id",
          "text",
          "score"
        ],
        "title": "RetrievalHit",
        "type": "object"
      },
      "RetrieveRequest": {
        "additionalProperties": false,
        "properties": {
          "documents": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "title": "Documents",
            "type": "array"
          },
          "k": {
            "default": 5,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          },
          "min_score": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Score",
            "type": "number"
          },
          "query": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "workspace_id": {
            "default": "default",
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "RetrieveRequest",
        "type": "object"
      },
      "RetrieveResponse": {
        "additionalProperties": false,
        "properties": {
          "hits": {
            "items": {
              "$ref": "#/components/schemas/RetrievalHit"
            },
            "title": "Hits",
            "type": "array"
          },
          "store": {
            "title": "Store",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "store",
          "workspace_id"
        ],
        "title": "RetrieveResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "Workspace": {
        "additionalProperties": false,
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/app__intelligence__context__store__StoredDocument"
            },
            "title": "Documents",
            "type": "array"
          },
          "store": {
            "title": "Store",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "store"
        ],
        "title": "Workspace",
        "type": "object"
      },
...
      "app__intelligence__context__store__StoredDocument": {
        "additionalProperties": false,
        "description": "A document as a workspace holds it.",
        "properties": {
          "chunks": {
            "default": 0,
            "title": "Chunks",
            "type": "integer"
          },
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "kind": {
            "default": "text",
            "title": "Kind",
            "type": "string"
          },
          "parser": {
            "default": "",
            "title": "Parser",
            "type": "string"
          },
          "sha256": {
            "default": "",
            "title": "Sha256",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "document_id"
        ],
        "title": "StoredDocument",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "An adaptive, evidence-driven investigation engine.\n\nMissions run in the background: `POST /api/v1/missions` returns a run id immediately and\n`GET /api/v1/missions/{id}/stream` streams the run as it happens. Every finding carries the\nsource locators it rests on, and every claim the evidence does not support is reported as\nunsupported rather than dropped.",
    "title": "JARVIS",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/context/ingest": {
      "post": {
        "description": "Chunk, embed and store documents in a workspace. Unchanged files are not re-embedded.",
        "operationId": "ingest_api_v1_context_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest",
        "tags": [
          "context"
        ]
      }
    },
    "/api/v1/context/retrieve": {
      "post": {
        "description": "The workspace's passages closest in meaning to the query, each a citation.",
        "operationId": "retrieve_api_v1_context_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retrieve",
        "tags": [
          "context"
        ]
      }
    },
...
    "/api/v1/context/workspaces/{workspace_id}": {
      "get": {
        "description": "A workspace's documents, with their parser, hash and chunk count.",
        "operationId": "workspace_api_v1_context_workspaces__workspace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workspace",
        "tags": [
          "context"
        ]
      }
    },
...
  }
}